requirements.txt
Flask==2.2.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.0
Werkzeug==2.2.3
Plotly==5.14.1
requests==2.28.2